redis 大改

+++ src/main/java/com/example/demo/control/CommentController.java
@@ -7,10 +7,14 @@ import com.example.demo.model.Paging;
 import com.example.demo.model.Result;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -18,11 +22,17 @@ public class CommentController {
 
     @Autowired
     private CommentDAO commentDAO;
+    @Autowired
+    private RedisTemplate redisTemplate;
 
     @ResponseBody
     @PostMapping("/comment")
     public CommentDO insert(@RequestBody CommentDO commentDO) {
+        if (commentDO == null) {
+            return null;
+        }
         commentDAO.insert(commentDO);
+        redisTemplate.opsForValue().set(commentDO.getId(),commentDO);
         return commentDO;
     }
 
@@ -30,13 +40,31 @@ public class CommentController {
     @ResponseBody
     @PostMapping("/comment/update")
     public CommentDO update(@RequestBody CommentDO commentDO) {
+        if (commentDO == null) {
+            return null;
+        }
         commentDAO.update(commentDO);
+        redisTemplate.opsForValue().set(commentDO.getId(), commentDO);
         return commentDO;
     }
 
     @ResponseBody
     @GetMapping("/comment/del")
     public boolean delete(@RequestParam Long id) {
+        if (id == null || id < 0) {
+            return false;
+        }
+        CommentDO commentDO = (CommentDO) redisTemplate.opsForValue().get(id);
+        if (commentDO == null) {
+            List<Long> ids = new ArrayList<>();
+            ids.add(id);
+            List<CommentDO> commentDOS = commentDAO.findByIds(ids);
+            if (CollectionUtils.isEmpty(commentDOS)) {
+                return false;
+            }
+            commentDO = commentDOS.get(0);
+        }
+        redisTemplate.delete(commentDO.getId());
         return commentDAO.delete(id) > 0;
     }
 
@@ -60,6 +88,10 @@ public class CommentController {
     @ResponseBody
     @GetMapping("/comment/findByRefId")
     public List<Comment> findByRefId(@RequestParam String refId) {
+        if (StringUtils.isBlank(refId)) {
+            return null;
+        }
+
         return commentDAO.findByRefId(refId);
     }
 
@@ -67,13 +99,22 @@ public class CommentController {
     @ResponseBody
     @PostMapping("/comment/batchAdd")
     public List<CommentDO> batchAdd(@RequestBody List<CommentDO> commentDOS) {
+        if (CollectionUtils.isEmpty(commentDOS)) {
+            return null;
+        }
+
         commentDAO.batchAdd(commentDOS);
+        commentDOS.forEach(commentDO -> redisTemplate.opsForValue().set(commentDO.getId(), commentDO));
         return commentDOS;
     }
 
     @ResponseBody
     @GetMapping("/comment/findByIds")
     public List<CommentDO> findByIds(@RequestParam List<Long> ids) {
+        if (CollectionUtils.isEmpty(ids)) {
+            return null;
+        }
+
         return commentDAO.findByIds(ids);
     }
